requirements.txt
pandas>=2.0
openpyxl>=3.1
psycopg2-binary>=2.9
requests>=2.31
beautifulsoup4>=4.12
rich>=13.0
python-dotenv>=1.0
click>=8.0
